break out app specific logic into app.ts, keep start function/process in index.ts
